feat(storybook-block):move utilities to specific util
