+++ packages/storybook-block/src/StorybookBlock.spec.ct.tsx
@@ -4,8 +4,7 @@ import { mount } from '@cypress/react';
 import { withAppBridgeBlockStubs } from '@frontify/app-bridge';
 import { ERROR_MSG } from './settings';
 import { StorybookBlock } from './StorybookBlock';
-import { StorybookBorderStyle, StorybookHeight, StorybookPosition, StorybookStyle, heights } from './types';
-import { decodeEntities } from './utilities';
+import { heights, StorybookBorderStyle, StorybookHeight, StorybookPosition, StorybookStyle } from './types';
 
 const StorybookBlockSelector = '[data-test-id="storybook-block"]';
 const EmptyStateSelector = '[data-test-id="storybook-empty-wrapper"]';
@@ -89,24 +88,5 @@ describe('Storybook Block', () => {
         cy.get(EmptyStateSelector);
         cy.get(EmptyStateSelector).find('button').should('be.disabled');
         cy.get(EmptyStateSelector).contains(ERROR_MSG);
-    });
-
-    describe('decodeEntities', () => {
-        it('decodes URL', () => {
-            const encodedUrl = 'https://fondue-components.frontify.com/?path&#61;/story/tokens--alias-tokens';
-            const decodedUrl = 'https://fondue-components.frontify.com/?path=/story/tokens--alias-tokens';
-            expect(decodedUrl).to.equal(decodeEntities(encodedUrl));
-        });
-
-        it('does nothing with already decoded URL', () => {
-            const decodedUrl = 'https://fondue-components.frontify.com/?path=/story/tokens--alias-tokens';
-            expect(decodedUrl).to.equal(decodeEntities(decodedUrl));
-        });
-
-        it('decodes characters', () => {
-            const encodedChars = '&#162;this&#38;&#174;&#60;is&#62;&#38;&#34&#39;awesome&#162;&#163;';
-            const decodedChats = '¢this&®<is>&"\'awesome¢£';
-            expect(decodedChats).to.equal(decodeEntities(encodedChars));
-        });
     });
 });
